feat(ilp): Refactor network job
